chore: update project structure
