+++ blueledger-webapp/components/nav-bar/desktop-nav-bar/index.ts
@@ -1,0 +1,1 @@
+export { default as DesktopNavBar } from './desktop-nav-bar';
