+++ blueledger-webapp/components/nav-bar/mobile-nav-bar/index.ts
@@ -1,0 +1,1 @@
+export { default as MobileNavBar } from './mobile-nav-bar';
